% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{check_functions}
\alias{check_functions}
\title{Heuristic functions checker}
\usage{
check_functions(file, loops = 0)
}
\arguments{
\item{file}{File name in working directory or path to file. File ('.txt') following the exact format from Atenea.}

\item{loops}{Number of loops allowed for the current heuristic, by default \emph{loops = 0}. See Details.}
}
\value{
Report with information about your result.
}
\description{
Self-check of allowed functions for all heuristics of Airport operations.
}
\details{
If any ERROR found, code will not run on any self-evaluator.

Maximum number of loops per activity:
\itemize{
\item Activity 2: 2 loops
\item Activity 4 (part 1): 0 loops
\item Activity 4 (part 2): 2 loops
}
}
\examples{
check_functions(file = "rs_heur.txt", loops = 2)
check_functions(file = "stand_util.txt", loops = 0)
check_functions(file = "sa_heur.txt", loops = 2)
}
