% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_eval.R
\name{rse_heur}
\alias{rse_heur}
\title{Runway Sequencing Heuristic Evaluation}
\usage{
rse_heur(file, toys_to_test = 1:10, recat_csv, icao_csv, group, get_grade)
}
\arguments{
\item{file}{File name in working directory or path to file. File ('rs_heur.txt') following the exact format from Atenea.}

\item{toys_to_test}{A vector with toy numbers to check your heuristic against. Database has 10 toy models and by default it runs all of them.}

\item{recat_csv}{File name in working directory or path to file. File ('rs_recat.csv') following the exact format from Atenea. Result of your heuristic using RECAT wake turbulence on your group's schedule.}

\item{icao_csv}{File name in working directory or path to file. File ('rs_icao.csv') following the exact format from Atenea. Result of your heuristic using ICAO wake turbulence on your group's schedule.}

\item{group}{Number of your group.}
}
\value{
Report with information about your result.
}
\description{
Self-evaluator for Activity 2 - RS Heuristic of Airport operations.
}
\details{
Maximum execution time of your code is 10 min.

Inputs \emph{recat_csv}, \emph{icao_csv} and \emph{group} are optional just for checking your schedule results. \emph{group} is mandatory when having a \emph{recat_csv} or \emph{icao_csv} input.
}
\examples{
rse_heur(file = "rs_heur.txt", toys_to_test = 1:10)
rse_heur(file = "rs_heur.txt", toys_to_test = 1:2, recat_csv = "rs_recat.csv", icao_csv = "rs_icao.csv", group = 1)
}
