% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_eval.R
\name{rse_manual}
\alias{rse_manual}
\title{Runway Sequencing Manual Evaluation}
\usage{
rse_manual(file, group)
}
\arguments{
\item{file}{File name in working directory or path to file. File ('.csv') following the exact format from Atenea.}

\item{group}{Number of your group.}
}
\value{
Report with information about your result.
}
\description{
Self-evaluator for Activity 1 - RS Manual of Airport operations.
}
\examples{
rse_manual(file = "rst_result.csv", group = 1)
}
