% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_eval.R
\name{sae_heur2}
\alias{sae_heur2}
\title{Stand Allocation Heuristic (part 2) Evaluation}
\usage{
sae_heur2(file, sch_to_test = 1:10, sa_csv, group, get_grade)
}
\arguments{
\item{file}{File name in working directory or path to file. File ('sa_heur.txt') following the exact format from Atenea.}

\item{sch_to_test}{A vector with schedule numbers to check your heuristic against. Database has schedules and by default it runs all of them.}

\item{sa_csv}{File name in working directory or path to file. File ('sa_heur.csv') following the exact format from Atenea. Result of your heuristic on your group's schedule.}

\item{group}{Number of your group.}
}
\value{
Report with information about your result.
}
\description{
Self-evaluator for Activity 4 - SA Heuristic (part 2) of Airport operations.
}
\details{
Maximum execution time of your code is 10 min.

Input \emph{sa_csv} and \emph{group} are optional just for checking your schedule results. \emph{group} is mandatory when having a \emph{sa_csv} input.
}
\examples{
sae_heur2(file = "sa_heur.txt", sch_to_test = c(1, 3))
sae_heur2(file = "sa_heur.txt", sch_to_test = 1:2, sa_csv = "sa_heur.csv", group = 1)
}
