% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_eval.R
\name{sae_manual}
\alias{sae_manual}
\title{Stand Allocation Manual Evaluation}
\usage{
sae_manual(file, group)
}
\arguments{
\item{file}{File name in working directory or path to file. File ('.csv') following the exact format from Atenea.}

\item{group}{Number of your group.}
}
\value{
Report with information about your result.
}
\description{
Self-evaluator for Activity 3 - SA Manual of Airport operations.
}
\examples{
rse_manual(file = "sa_result.csv", group = 1)
}
